# web/docs/swagger.yaml
definitions:
...
  models.ResUser:
    properties:
      session:
        type: string
      user:
        $ref: '#/definitions/models.ResUserInfo'
    type: object
  models.ResUserInfo:
    properties:
      agent_balance:
        description: 代理余额
        type: integer
      balance:
        type: integer
      email:
        type: string
      flow:
        description: 用户剩余流量 GB
        type: string
      flow_date:
        description: 用户剩余流量到期时间
        type: string
      flow_expire:
        description: 用户剩余流量是否到期
        type: integer
      flow_mb:
        description: 用户剩余流量 MB
        type: string
      flow_mb_unit:
        description: 用户剩余流量单位 MB
        type: string
      flow_redeem:
        description: 可兑换 流量余额 GB
        type: string
      flow_unit:
        description: 用户剩余流量单位 GB
        type: string
      google_auth:
        description: 是否 已经绑定  1 已经绑定
        type: integer
      id:
        type: integer
      invite:
        description: 邀请码
        type: string
      invite_url:
        description: 邀请地址
        type: string
      ip_num:
        description: 剩余IP数
        type: integer
      is_act:
        description: 用户剩余流量是否到期
        type: integer
      is_pay:
        description: 是否是已购买
        type: integer
      nickname:
        type: string
      username:
        type: string
    type: object
...
host: textapi.360proxy.net
info:
  contact: {}
  description: This 360 API
  title: 360 API
  version: "1.0"
paths:
...
  /web/user/auto_login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 自动登录
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            items:
              $ref: '#/definitions/models.ResUser'
            type: array
      summary: 自动登录
      tags:
      - 登陆注册相关
  /web/user/bind_email:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 绑定邮箱
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            type: object
      summary: 绑定邮箱
      tags:
      - 登陆注册相关
  /web/user/bind_wallet:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 绑定钱包地址
      parameters:
      - description: 钱包地址
        in: formData
        name: wallet
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      - description: 竞价平台
        in: formData
        name: j_source
        type: string
      - description: 竞价code
        in: formData
        name: j_code
        type: string
      - description: 竞价域名
        in: formData
        name: j_domain
        type: string
      - description: 终端
        in: formData
        name: p
        type: string
      - description: 来源
        in: formData
        name: origin
        type: string
      - description: 邀请类型 1链接  2填写推广码
        in: formData
        name: inviter_type
        type: string
      - description: 邀请码
        in: formData
        name: inviter_code
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: wallet：钱包地址
          schema:
            additionalProperties: true
            type: object
      summary: 绑定钱包地址
      tags:
      - 个人中心
...
  /web/user/email_info:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 根据邮箱获取用户信息
      parameters:
      - description: 邮箱
        in: formData
        name: email
        type: string
      - description: 邀请码
        in: formData
        name: invite
        type: string
      - description: 套餐ID
        in: formData
        name: package_id
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            additionalProperties: true
            type: object
      summary: 根据邮箱获取用户信息
      tags:
      - 个人中心
...
  /web/user/pay_list:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 用户购买记录
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 每页数据量
        in: formData
        name: limit
        type: string
      - description: 页数
        in: formData
        name: page
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 'total:总记录数,total_page:总页数,lists:记录列表 '
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: 用户购买记录
      tags:
      - 个人中心
  /web/user/reset_pwd:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 修改 /重置密码
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 旧密码
        in: formData
        name: old_password
        required: true
        type: string
      - description: 新密码
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: re_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            type: object
      summary: 修改 /重置密码
      tags:
      - 登陆注册相关
...
  /web/user/unbind_wallet:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 解绑钱包地址
      parameters:
      - description: 邮箱
        in: formData
        name: email
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            type: object
      summary: 解绑钱包地址
      tags:
      - 个人中心
...
  /web/user/used_data:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 用户IP消耗日期和数量
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 开始时间
        in: formData
        name: start_date
        type: string
      - description: 结束时间
        in: formData
        name: end_date
        type: string
      - description: 类型 isp、static、flow
        in: formData
        name: cate
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 'x_data : x轴数据, y_data : y轴数据, cate :分类名称, unit :单位, max:最大值'
          schema:
            additionalProperties: true
            type: object
      summary: 用户IP消耗日期和数量
      tags:
      - 个人中心
  /web/user/used_ip_data:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 用户IP消耗日期和数量
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 'x_data : x轴数据, y_data: y轴数据, cate:分类名称'
          schema:
            additionalProperties: true
            type: object
      summary: 用户IP消耗日期和数量
      tags:
      - 个人中心
  /web/user/used_ip_list:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 用户IP消耗记录
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 每页数据量
        in: formData
        name: limit
        type: string
      - description: 页数
        in: formData
        name: page
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 'total:总记录数,total_page:总页数,lists:记录列表 '
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: 用户IP消耗记录
      tags:
      - 个人中心
...
  /web/user/user_info:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取用户信息
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            items:
              $ref: '#/definitions/models.ResUser'
            type: array
      summary: 获取用户信息
      tags:
      - 个人中心
...
schemes:
- http
- https
swagger: "2.0"